% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Add a \code{label} attribute to a vector}
\usage{
add_label(data, label)
}
\arguments{
\item{data}{Vector-like object.}

\item{label}{Label text.}
}
\value{
\code{data} with a \code{label} attribute.
}
\description{
Add a \code{label} attribute to a vector
}
\examples{
add_label(1:3, "Example variable")
}
