% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{codebook_strings_to_ignore}
\alias{codebook_strings_to_ignore}
\title{Codebook strings to ignore, as it contains HTML formatting}
\format{
A data frame.
}
\usage{
codebook_strings_to_ignore
}
\description{
Codebook strings to ignore, as it contains HTML formatting
}
\keyword{datasets}
