% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{melidos_cities}
\alias{melidos_cities}
\title{MeLiDos site cities}
\format{
A named character vector.
}
\source{
MeLiDos study metadata.
}
\usage{
melidos_cities
}
\description{
Named character vector of city names keyed by study site abbreviation.
}
\keyword{datasets}
