% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{melidos_colors}
\alias{melidos_colors}
\title{MeLiDos site colors}
\format{
A named character vector.
}
\source{
MeLiDos study metadata.
}
\usage{
melidos_colors
}
\description{
Named character vector of hex colors keyed by study site abbreviation.
}
\keyword{datasets}
