% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{melidos_coordinates}
\alias{melidos_coordinates}
\title{MeLiDos site coordinates}
\format{
A named list.
}
\source{
MeLiDos study metadata.
}
\usage{
melidos_coordinates
}
\description{
Named list of latitude/longitude numeric pairs keyed by study site
abbreviation.
}
\keyword{datasets}
