% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{melidos_countries}
\alias{melidos_countries}
\title{MeLiDos site countries}
\format{
A named character vector.
}
\source{
MeLiDos study metadata.
}
\usage{
melidos_countries
}
\description{
Named character vector of country names keyed by study site abbreviation.
}
\keyword{datasets}
