% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{melidos_tzs}
\alias{melidos_tzs}
\title{MeLiDos site time zones}
\format{
A named character vector.
}
\source{
MeLiDos study metadata.
}
\usage{
melidos_tzs
}
\description{
Named character vector of Olson time zone identifiers keyed by study site
abbreviation.
}
\keyword{datasets}
