% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melidosData-package.R
\docType{package}
\name{melidosData-package}
\alias{melidosData}
\alias{melidosData-package}
\title{melidosData: Utilities and example assets for MeLiDos data workflows}
\description{
The package includes helper functions for working with REDCap codebooks and
records, as well as small example datasets and dictionaries for reproducible
examples.
}
\author{
\strong{Maintainer}: Johannes Zauner \email{johannes.zauner@tum.de}

Other contributors:
\itemize{
  \item European Partnership on Metrology (The project (22NRM05 MeLiDos) has received funding from the European Partnership on Metrology, co-financed by the European Union's Horizon Europe Research and Innovation Programme, EURAMET, and the Participating States. Views and opinions expressed are those of the authors and do not necessarily reflect those of the European Union or EURAMET.) [funder]
  \item Translational Sensory and Circadian Neuroscience Unit (MPS/TUM/TUMCREATE) (www.tscnlab.org) [copyright holder]
}

}
\keyword{internal}
