% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_codebook_prepare.R
\name{REDCap_codebook_prepare}
\alias{REDCap_codebook_prepare}
\title{Prepare a REDCap codebook for downstream processing}
\usage{
REDCap_codebook_prepare(
  codebook,
  strings_to_ignore = codebook_strings_to_ignore,
  form.filter = NULL,
  field.label = `Field Label`,
  form.col = `Form Name`
)
}
\arguments{
\item{codebook}{A REDCap data dictionary as a data frame.}

\item{strings_to_ignore}{A regular expression used in
\code{\link[stringr:str_remove]{stringr::str_remove_all()}} to remove formatting fragments from field labels.}

\item{form.filter}{Optional character vector of REDCap form names to keep.}

\item{field.label}{Unquoted column containing question labels.}

\item{form.col}{Unquoted column containing the form name (for filtering)}
}
\value{
A cleaned codebook tibble/data frame.
}
\description{
Cleans HTML formatting from codebook labels and optionally filters the
codebook to selected REDCap forms.
}
\examples{
codebook_path <- system.file("ext", "DataDictionary_sleepdiary.csv",
  package = "melidosData"
)
codebook <- utils::read.csv(codebook_path, check.names = FALSE)
cleaned <- REDCap_codebook_prepare(codebook)
cleaned$`Field Label`
#compate the original one
codebook$`Field Label`
}
