% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_labels.R
\name{REDCap_col_labels}
\alias{REDCap_col_labels}
\title{Add variable labels from a REDCap codebook}
\usage{
REDCap_col_labels(
  data,
  lookup,
  var_col = `Variable / Field Name`,
  label_col = `Field Label`,
  warn = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{lookup}{A data frame with variable names and labels.}

\item{var_col}{Unquoted column in \code{lookup} containing variable names.}

\item{label_col}{Unquoted column in \code{lookup} containing human-readable
labels.}

\item{warn}{Logical; warn when labels are provided for variables not present
in \code{data}.}
}
\value{
\code{data} with \code{label} attributes added.
}
\description{
Sets the \code{label} attribute on matching columns in \code{data}.
}
\examples{
dict_path <- system.file("ext", "DataDictionary_chronotype.csv",
  package = "melidosData"
)
dict <- utils::read.csv(dict_path, check.names = FALSE)
dict <- REDCap_codebook_prepare(dict,  form.filter = "mctq")
labelled <- REDCap_col_labels(REDCap_example_chronotype, dict)
attr(labelled[[5]], "label")
}
