% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_coltype_check.R
\name{REDCap_coltype_check}
\alias{REDCap_coltype_check}
\title{Check data column classes against REDCap expectations}
\usage{
REDCap_coltype_check(
  codebook,
  indicator_POSIXct = "datetime_dmy",
  indicator_date = "Date",
  indicator_time = "time",
  indicator_logical = "yesno",
  indicator_numeric.val_col = c("number", "integer"),
  indicator_numeric.type_col = c("radio", "dropdown"),
  label_col = `Field Label`,
  name_col = `Variable / Field Name`,
  type_col = `Field Type`,
  val_col = `Text Validation Type OR Show Slider Number`,
  data
)
}
\arguments{
\item{codebook}{REDCap data dictionary.}

\item{indicator_POSIXct}{Indicator in \code{val_col} identifying datetime fields.}

\item{indicator_date}{Pattern used in labels to identify date variables.}

\item{indicator_time}{Indicator in \code{val_col} identifying time-only fields.}

\item{indicator_logical}{Indicator in \code{type_col} identifying logical fields.}

\item{indicator_numeric.val_col}{Indicators in \code{val_col} for numeric fields.}

\item{indicator_numeric.type_col}{Indicators in \code{type_col} for numeric fields.}

\item{label_col}{Unquoted codebook label column.}

\item{name_col}{Unquoted codebook variable-name column.}

\item{type_col}{Unquoted codebook field-type column.}

\item{val_col}{Unquoted codebook validation/type-hint column.}

\item{data}{Data frame to validate.}
}
\value{
A list with \code{ok}, \code{summary}, and per-column \code{details}.
}
\description{
Uses REDCap codebook metadata to infer expected classes and compares these to
classes in \code{data}.
}
\examples{

library(gt)
dict_path <- system.file("ext", "DataDictionary_sleepdiary.csv",
  package = "melidosData"
)
dict <- utils::read.csv(dict_path, check.names = FALSE)

coltype_check <- REDCap_coltype_check(dict, data = REDCap_example_sleep)
coltype_check$ok
coltype_check$summary
coltype_check$details |> gt()

dict_path <- system.file("ext", "DataDictionary_chronotype.csv",
  package = "melidosData"
)
dict <- utils::read.csv(dict_path, check.names = FALSE)
dict <- REDCap_codebook_prepare(dict, form.filter = "mctq")

coltype_check <- REDCap_coltype_check(dict, data = REDCap_example_chronotype)
coltype_check$ok
coltype_check$summary
coltype_check$details
}
