% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{REDCap_example_chronotype}
\alias{REDCap_example_chronotype}
\title{Example REDCap chronotype data}
\format{
A data frame.
}
\source{
\code{inst/ext/example_chronotype.csv}.
}
\usage{
REDCap_example_chronotype
}
\description{
Example export containing chronotype questionnaire responses.
}
\keyword{datasets}
