% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{REDCap_example_sleep}
\alias{REDCap_example_sleep}
\title{Example REDCap sleep diary data}
\format{
A data frame.
}
\source{
\code{inst/ext/example_sleepdiary.csv}.
}
\usage{
REDCap_example_sleep
}
\description{
Example export containing sleep diary responses.
}
\keyword{datasets}
