% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melidosData-package.R
\docType{package}
\name{melidosData-package}
\alias{melidosData}
\alias{melidosData-package}
\title{melidosData: Utilities and example assets for MeLiDos data workflows}
\description{
The package includes helper functions for working with REDCap codebooks and
records, as well as small example datasets and dictionaries for reproducible
examples.
}
\author{
\strong{Maintainer}: Johannes Zauner \email{johannes.zauner@tum.de}

}
\keyword{internal}
